simplify demo file structure
