+++ src/PostgresOutbox/Serialization/JsonSerialization.cs
@@ -1,4 +1,3 @@
-using System.Data.Common;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Text.Json.Serialization.Metadata;
