--- src/Publisher/SourceGenerationContext.cs
@@ -1,10 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace Publisher;
-
-[JsonSourceGenerationOptions(WriteIndented = true)]
-[JsonSerializable(typeof(UserCreated))]
-[JsonSerializable(typeof(UserDeleted))]
-internal partial class SourceGenerationContext: JsonSerializerContext
-{
-}
